Starting maktaboo
Started writing maktaboo as a server-client architecture with data
shared by sockets. The overarching plan is that once the maktaboo
server is good enough, we can implement its client in multiple editors
and allow cool bibliography management from these editors
directly. For the CLI client, I chose to use the python libraries click and
questionary to get a better interface.